refactor: distinguish Promise and stream
