+++ src/mapToObject.js
