+++ src/timedRun.js
@@ -9,9 +9,15 @@ const juration = require('juration');
 * @private
 * @param {!Object} logger A {@link logger} object
 * @param {!Function} runFn The function to run
+* @return {Function} a function to start the processing. Takes an optional callback parameter.
 */
 module.exports = (logger, runFn) => {
-    new NanoTimer().time(runFn, '', 's', (time) => {
-        logger.info(`Finished. Took ${juration.stringify(time) || 'less than a second'}.`);
-    });
+    return (cb) => {
+        new NanoTimer().time(runFn, '', 's', (time) => {
+            logger.info(`Finished. Took ${juration.stringify(time) || 'less than a second'}.`);
+            if (typeof cb === 'function') {
+                cb();
+            }
+        });
+    };
 };
